move version and inspect api testcase to passed
